make CAPI api call from callsite and remove now unused FacebookCAPI struct
