--- Library/Tracking/FacebookCAPITests.swift
@@ -1,26 +1,0 @@
-@testable import KsApi
-@testable import Library
-import XCTest
-
-final class FacebookCAPITests: XCTestCase {
-  func testCreateMutationInput() {
-    let mutationInput = FacebookCAPI
-      .createMutationInput(
-        for: .ProjectPageViewed,
-        projectId: "projId",
-        externalId: "extId",
-        userEmail: "email",
-        currency: "currency",
-        value: "value"
-      )
-
-    let dict = mutationInput.toInputDictionary()
-
-    XCTAssertEqual(dict["eventName"] as! String, FacebookCAPIEventName.ProjectPageViewed.rawValue)
-    XCTAssertEqual(dict["projectId"] as! String, "projId")
-    XCTAssertEqual(dict["externalId"] as! String, "extId")
-    XCTAssertEqual(dict["userEmail"] as! String, "email")
-    XCTAssertNotNil(dict["appData"])
-    XCTAssertNotNil(dict["customData"])
-  }
-}
